+++ app/components/flowbite/input_field/radio_button.rb
@@ -26,6 +26,15 @@ module Flowbite
         input_component.new(**args)
       end
 
+      def default_hint_options
+        return {} unless @hint
+
+        {
+          class: hint_classes,
+          id: id_for_hint_element
+        }.merge(@hint[:options] || {})
+      end
+
       # Returns options for the default label element. This includes CSS classes
       # since they are specific to RadioButton labels (and Checkbox ones).
       def default_label_options
@@ -35,21 +44,6 @@ module Flowbite
             for: id_for_input_element
           }
         })
-      end
-
-      # Returns the HTML to use for the hint element if any
-      def hint
-        return unless hint?
-
-        component = Flowbite::Input::Hint.new(
-          attribute: @attribute,
-          form: @form,
-          options: {
-            class: hint_classes,
-            id: id_for_hint_element
-          }
-        ).with_content(@hint)
-        render(component)
       end
 
       def input_component