+++ demo/test/components/previews/radio_button_preview.rb
@@ -50,7 +50,7 @@ class RadioButtonPreview < Lookbook::Preview
   # the checkbox component.
 
   def checkbox_with_helper_text
-    render(Flowbite::InputField::RadioButton.new(attribute: :state, form: form, value: "visible", hint: "I agree with the terms and conditions."))
+    render(Flowbite::InputField::RadioButton.new(attribute: :state, form: form, value: "visible", hint: {content: "I agree with the terms and conditions."}))
   end
 
   # @!endgroup